Remove populateUserRegion()
Enabled by our changes to UserService.getActingUser(), we remove populateUserRegion(), and move its responsibilities into populateCurrentUser(). This saves two largely identical tasks being spread across different pieces of middleware. We take this as an opportunity to add missing testing to populateCurrentUser().
