+++ server/middleware/populateCurrentUser.ts
@@ -9,7 +9,10 @@ export default function populateCurrentUser(userService: UserService): RequestHa
       if (req.user) {
         const callConfig = extractCallConfig(req)
         const userDetails = req.session.userDetails || (await userService.getActingUser(callConfig))
+
         req.session.userDetails = userDetails
+        req.session.probationRegion = userDetails.region
+
         res.locals.user = { ...userDetails, ...res.locals.user }
 
         if (!userDetails) {
