+++ server/middleware/setUpCurrentUser.ts
@@ -2,14 +2,11 @@ import { Router } from 'express'
 import auth from '../authentication/auth'
 import tokenVerifier from '../data/tokenVerification'
 import type { Services } from '../services'
-import populateUserRegion from './populateUserRegion'
-import asyncMiddleware from './asyncMiddleware'
 import populateCurrentUser from './populateCurrentUser'
 
 export default function setUpCurrentUser({ userService }: Services): Router {
   const router = Router({ mergeParams: true })
   router.use(auth.authenticationMiddleware(tokenVerifier))
   router.use(populateCurrentUser(userService))
-  router.use(asyncMiddleware(populateUserRegion(userService)))
   return router
 }
